+++ src/components/Point/Point.tsx
@@ -9,17 +9,40 @@ interface Props {
   cx: number;
   cy: number;
   color: Color;
+  index?: number;
+  onFocus?: ({index, cx, cy}: {index?: number; cx: number; cy: number}) => any;
+  tabIndex?: number;
+  ariaLabelledby?: string;
 }
 
-export function Point({cx, cy, active, color}: Props) {
+export function Point({
+  cx,
+  cy,
+  active,
+  color,
+  onFocus,
+  index,
+  ariaLabelledby,
+  tabIndex = -1,
+}: Props) {
+  const handleFocus = () => {
+    if (onFocus != null) {
+      onFocus({index, cx, cy});
+    }
+  };
+
   return (
     <circle
+      role={ariaLabelledby == null ? '' : 'image'}
+      aria-labelledby={ariaLabelledby}
+      tabIndex={tabIndex}
       cx={cx}
       cy={cy}
       r={active ? 5 : 0}
       fill={getColorValue(color)}
       stroke={tokens.colorWhite}
       strokeWidth={1.5}
+      onFocus={handleFocus}
     />
   );
 }